% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptairData-package.R
\docType{package}
\name{ptairData-package}
\alias{ptairData}
\alias{ptairData-package}
\title{ptairData: PTR-TOF-MS volatolomics raw datasets from exhaled air and cell culture headspace}
\description{
The package ptairData contains two raw datasets from Proton-Transfer-Reaction Time-of-Flight mass 
    spectrometer acquisitions (PTR-TOF-MS), in the HDF5 format: one from the exhaled air of two volunteer healthy 
    individuals, and one from the cell culture headspace from two mycobacteria species and one control 
    (culture medium only). Those datasets are used in the examples and in the vignette from the ptairMS package 
    (PTR-TOF-MS data pre-processing).
}
\author{
\strong{Maintainer}: camille Roquencourt \email{camille.roquencourt@hotmail.fr}

}
\keyword{internal}
